# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Family Manager BI Group API",
        "version": "1.1.0"
    },
    "paths": {
...
        "/api/case-versions": {
            "get": {
                "tags": [
                    "CaseVersions"
                ],
                "description": "Get list of cases with search",
                "operationId": "App\\Http\\Controllers\\CaseVersionController::index",
                "parameters": [
                    {
                        "name": "case_hash",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return sections list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CaseVersionResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/case-versions/{hash}": {
            "put": {
                "tags": [
                    "CaseVersions"
                ],
                "description": "Update status CaseVersion",
                "operationId": "App\\Http\\Controllers\\CaseVersionController::update",
                "parameters": [
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "CaseVersion hash",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CaseVersionStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category sucsessfully change"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CaseVersionStatusRequest": {
                "title": "Case version status request",
                "description": "Case version status request body data",
                "required": [
                    "is_active"
                ],
                "properties": {
                    "is_active": {
                        "schema": "CaseVersionStatusRequest",
                        "type": "boolean"
                    },
                    "case_id": {
                        "schema": "CaseVersionStatusRequest",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "CaseVersionResource": {
                "title": "CaseVersionSearchResource Resource",
                "description": "CaseFile resource",
                "properties": {
                    "name": {
                        "schema": "CaseVersionResource",
                        "description": "Class CaseVersionResource",
                        "type": "string"
                    },
                    "hash": {
                        "schema": "CaseVersionResource",
                        "description": "Class CaseVersionResource",
                        "type": "string"
                    },
                    "type_id": {
                        "schema": "CaseVersionResource",
                        "description": "Class CaseVersionResource",
                        "type": "integer"
                    },
                    "version": {
                        "schema": "CaseVersionResource",
                        "description": "Class CaseVersionResource",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}